+++ interface/internals/MapWidget.py
